+++ examples/vue/.yalc/@use-modals/vue/src/plugin.ts
@@ -1,9 +1,0 @@
-import Vue from 'vue'
-
-import { injectMethods } from './method'
-import { installComponent } from './component'
-
-export const install = (vm: typeof Vue) => {
-  injectMethods(vm)
-  installComponent(vm)
-}