+++ examples/vue3/.yalc/@use-modals/vue3/src/index.ts
@@ -1,5 +1,0 @@
-import { UseModals } from './plugin'
-
-export { useOpenModal, useUpdateModal, useCloseModal, useCloseAllModals } from './hooks'
-
-export default UseModals