+++ examples/vue3/.yalc/@use-modals/vue3/src/plugin.ts
@@ -1,9 +1,0 @@
-import { Plugin, App } from 'vue'
-
-import { installComponent } from './component'
-
-export const UseModals: Plugin = {
-  install(app: App) {
-    installComponent(app)
-  }
-}